requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bandit==1.8.3
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dishka==1.4.2
dnspython==2.7.0
email_validator==2.2.0
fast-depends==2.4.12
fastapi==0.115.8
fastapi-cli==0.0.7
faststream==0.5.34
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
orjson==3.10.15
packaging==24.2
pbr==6.1.1
pluggy==1.5.0
punq==0.7.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
stevedore==5.4.1
Telethon==1.39.0
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
opentelemetry-api==1.32.1 ; python_version >= "3.11"
opentelemetry-exporter-jaeger-proto-grpc==1.21.0 ; python_version >= "3.11"
opentelemetry-exporter-jaeger-thrift==1.21.0 ; python_version >= "3.11"
opentelemetry-exporter-jaeger==1.21.0 ; python_version >= "3.11"
opentelemetry-exporter-otlp-proto-common==1.32.1 ; python_version >= "3.11"
opentelemetry-exporter-otlp-proto-grpc==1.32.1 ; python_version >= "3.11"
opentelemetry-exporter-otlp-proto-http==1.32.1 ; python_version >= "3.11"
opentelemetry-exporter-otlp==1.32.1 ; python_version >= "3.11"
opentelemetry-instrumentation-asgi==0.53b1 ; python_version >= "3.11"
opentelemetry-instrumentation-fastapi==0.53b1 ; python_version >= "3.11"
opentelemetry-instrumentation==0.53b1 ; python_version >= "3.11"
opentelemetry-proto==1.32.1 ; python_version >= "3.11"
opentelemetry-sdk==1.32.1 ; python_version >= "3.11"
opentelemetry-semantic-conventions==0.53b1 ; python_version >= "3.11"
opentelemetry-util-http==0.53b1 ; python_version >= "3.11"